# huff0: Use bmi1 on GOAMD64=v3