--- a/huff0/decompress_amd64.s.in
+++ b/huff0/decompress_amd64.s.in
@@ -6,6 +6,11 @@
 #include "funcdata.h"
 #include "go_asm.h"
 
+#ifdef GOAMD64_v4
+#ifndef GOAMD64_v3
+#define GOAMD64_v3
+#endif
+#endif
 
 #define bufoff      256     // see decompress.go, we're using [4][256]byte table
 
@@ -49,9 +54,9 @@ main_loop:
     MOVQ    bitReaderShifted_value(br{{ var "id" }}), br_value
     MOVQ    bitReaderShifted_off(br{{ var "id" }}), br_offset
 
-	// if b.bitsRead >= 32 {
-    CMPQ    br_bits_read, $32
-    JB      skip_fill{{ var "id" }}
+    // We must have at least 2 * max tablelog left
+    CMPQ    br_bits_read, $64-22
+    JBE     skip_fill{{ var "id" }}
 
     SUBQ    $32, br_bits_read       // b.bitsRead -= 32
     SUBQ    $4, br_offset           // b.off -= 4
@@ -60,11 +65,16 @@ main_loop:
 	// v = v[:4]
 	// low := (uint32(v[0])) | (uint32(v[1]) << 8) | (uint32(v[2]) << 16) | (uint32(v[3]) << 24)
     MOVQ    bitReaderShifted_in(br{{ var "id" }}), AX
-    MOVL    0(br_offset)(AX*1), AX  // AX = uint32(b.in[b.off:b.off+4])
 
 	// b.value |= uint64(low) << (b.bitsRead & 63)
+#ifdef GOAMD64_v3
+    SHLXQ   br_bits_read, 0(br_offset)(AX*1), AX // AX = uint32(b.in[b.off:b.off+4]) << (b.bitsRead & 63)
+#else
+    MOVL    0(br_offset)(AX*1), AX  // AX = uint32(b.in[b.off:b.off+4])
     MOVQ    br_bits_read, CX
     SHLQ    CL, AX
+#endif
+
     ORQ     AX, br_value
 
     // exhausted = exhausted || (br{{ var "id"}}.off < 4)
@@ -75,31 +85,54 @@ main_loop:
 skip_fill{{ var "id" }}:
 
     // val0 := br{{ var "id"}}.peekTopBits(peekBits)
+#ifdef GOAMD64_v3
+    SHRXQ   peek_bits, br_value, AX // AX = (value >> peek_bits) & mask
+#else
     MOVQ    br_value, AX
     MOVQ    peek_bits, CX
     SHRQ    CL, AX                  // AX = (value >> peek_bits) & mask
+#endif
 
     // v0 := table[val0&mask]
     MOVW    0(table)(AX*2), AX      // AX - v0
 
     // br{{ var "id"}}.advance(uint8(v0.entry))
     MOVB    AH, BL                  // BL = uint8(v0.entry >> 8)
+
+#ifdef GOAMD64_v3
+    MOVBQZX AL, CX
+    SHLXQ   AX, br_value, br_value // value <<= n
+#else
     MOVBQZX AL, CX
     SHLQ    CL, br_value            // value <<= n
+#endif
+
     ADDQ    CX, br_bits_read        // bits_read += n
 
+
+#ifdef GOAMD64_v3
+    SHRXQ    peek_bits, br_value, AX  // AX = (value >> peek_bits) & mask
+#else
     // val1 := br{{ var "id"}}.peekTopBits(peekBits)
     MOVQ    peek_bits, CX
     MOVQ    br_value, AX
     SHRQ    CL, AX                  // AX = (value >> peek_bits) & mask
+#endif
 
     // v1 := table[val1&mask]
     MOVW    0(table)(AX*2), AX      // AX - v1
 
     // br{{ var "id"}}.advance(uint8(v1.entry))
     MOVB    AH, BH                  // BH = uint8(v1.entry >> 8)
+
+#ifdef GOAMD64_v3
     MOVBQZX AL, CX
-    SHLQ    CX, br_value            // value <<= n
+    SHLXQ   AX, br_value, br_value // value <<= n
+#else
+    MOVBQZX AL, CX
+    SHLQ    CL, br_value            // value <<= n
+#endif
+
     ADDQ    CX, br_bits_read        // bits_read += n