@klauspost 2022-03-08
@WojciechMula For tiny tweaks like this, we can use the conditional compilation added in v1.18

I doubt it will be measurable in the final output, so this is an easy for people to get out the last performance.