@WojciechMula 2022-03-10
> @WojciechMula For tiny tweaks like this, we can use the conditional compilation added in v1.18
> 
> I doubt it will be measurable in the final output, so this is an easy for people to get out the last performance.
> 
> I don't know if you've seen it, but setting `GOAMD64=v3` will enable "v3" features. See list of what that includes here: https://en.wikipedia.org/wiki/X86-64#Microarchitecture_levels

Didn't know about that feature, thanks for the info. However, I am more for runtime dispatching, especially that in Go we don't have to deal with multiple calling conventions (in C/C++ it's pain in the neck). Additionally, my tests showed that using a function pointer doesn't affect on performance -- a branch predictor works well.