@klauspost 2022-03-10
@WojciechMula Yeah, see [s2 compression](https://github.com/klauspost/compress/blob/master/s2/encodeblock_amd64.s) if you want to see that. 

However, it is also a tradeoff. bmi is a good example in the s2 code. There are tiny improvements to be had with bmi, but it would double the amount of output assembler from 16k LOC to 32k. Given that the difference is so small I just went for the conditional compilation.

I don't know if you looked into it, but I feel like using `avo` has significant productivity improvements compared to text templates. I generally feel that it makes your code easier to maintain and refactor, since it allows dynamic registers. But if you aren't comfortable with it, feel free to submit your suggestions as-is, and I will look into converting it.