@WojciechMula 2022-03-10
> I don't know if you looked into it, but I feel like using `avo` has significant productivity improvements compared to text templates. I generally feel that it makes your code easier to maintain and refactor, since it allows dynamic registers. But if you aren't comfortable with it, feel free to submit your suggestions as-is, and I will look into converting it.

I heard about `avo`, but never used it. Since I already have text templates, I'd like to commit them as-is and when everything will be stable, I'll convert them into `avo`. It shouldn't be your job. :)

Speaking of code versions: I'll go for a conditional compilation.